Split models into multiple files
Split the models python file into a module with multiple files
