+++ django/repository/models/__init__.py
@@ -1,0 +1,7 @@
+from .discord_bot import *
+from .package import *
+from .package_download import *
+from .package_rating import *
+from .package_version import *
+from .uploader_identity import *
+from .uploader_identity import *
